refactor: moved common env replacements in vite.utils.ts as per review
